+++ lib/reduced.dart
@@ -1,6 +1,7 @@
 // reduced.dart
 
-export 'src/reducible.dart';
+export 'src/reducer.dart';
+export 'src/store.dart';
 export 'src/callable.dart';
 export 'src/callbacks.dart';
 export 'src/functions.dart';