+++ lib/src/callable.dart
@@ -1,6 +1,7 @@
 // callable.dart
 
-import 'reducible.dart';
+import 'reducer.dart';
+import 'store.dart';
 
 /// An abstraction for callbacks without parameters in the form of a class to easily implement value semantics.
 ///
@@ -96,150 +97,151 @@ abstract class Callable3<R, V1, V2, V3> {
   R call(V1 value1, V2 value2, V3 value3);
 }
 
-/// An implementation of a callback as a [Reducible.reduce](Reducible.reduce) call with a [Reducer].
+/// An implementation of a callback as a [ReducedStore.reduce](ReducedStore.reduce) call with a [Reducer].
 ///
-/// Or in other words, a [Reducer] bonded to a [Reducible] useable as callback.
+/// Or in other words, a [Reducer] bonded to a [ReducedStore] useable as callback.
 /// Can be assigned to Widget properties of type [VoidCallback].
-/// The type parameter `S` is the type of the state of the [Reducible].
+/// The type parameter `S` is the type of the state of the [ReducedStore].
 class CallableAdapter<S> extends Callable<void> {
-  const CallableAdapter(this.reducible, this.reducer);
+  const CallableAdapter(this.store, this.reducer);
 
-  /// The reducible to whose method [reduce](Reducible.reduce)
+  /// The store to whose method [reduce](ReducedStore.reduce)
   /// the [reducer] is passed when the method [call] is called.
-  final Reducible<S> reducible;
+  final ReducedStore<S> store;
 
-  /// The reducer that is passed as a parameter to the [reduce](Reducible.reduce) method
-  /// of the [reducible] when the [call] method is called.
+  /// The reducer that is passed as a parameter to the [reduce](ReducedStore.reduce) method
+  /// of the [store] when the [call] method is called.
   final Reducer<S> reducer;
 
-  /// Executes the [reduce](Reducible.reduce) method of the [reducible]
+  /// Executes the [reduce](ReducedStore.reduce) method of the [store]
   /// with the [reducer] as parameter.
   @override
-  call() => reducible.reduce(reducer);
+  call() => store.reduce(reducer);
 
   /// For this class to have value semantics, both constructor parameters
-  /// [reducible] and [reducer] should have value semantics.
+  /// [store] and [reducer] should have value semantics.
   @override
-  get hashCode => Object.hash(reducible, reducer);
+  get hashCode => Object.hash(store, reducer);
 
   /// For this class to have value semantics, both constructor parameters
-  /// [reducible] and [reducer] should have value semantics.
+  /// [store] and [reducer] should have value semantics.
   @override
   operator ==(other) =>
       other is CallableAdapter &&
       reducer == other.reducer &&
-      reducible == other.reducible;
+      store == other.store;
 }
 
-/// An implementation of a callback as a [Reducible.reduce](Reducible.reduce) call with a [Reducer1].
+/// An implementation of a callback as a [ReducedStore.reduce](ReducedStore.reduce) call with a [Reducer1].
 ///
-/// Or in other words, a [Reducer] bonded to a [Reducible] useable as callback.
+/// Or in other words, a [Reducer] bonded to a [ReducedStore] useable as callback.
 ///
-/// The type parameter `S` is the type of the state of the [Reducible].
+/// The type parameter `S` is the type of the state of the [ReducedStore].
 /// The type parameter `V` is the type of the value of the [Reducer1].
 class Callable1Adapter<S, V> extends Callable1<void, V> {
-  const Callable1Adapter(this.reducible, this.reducer);
+  const Callable1Adapter(this.store, this.reducer);
 
-  /// The reducible to whose method [reduce](Reducible.reduce)
+  /// The store to whose method [reduce](ReducedStore.reduce)
   /// the [reducer] is passed when the method [call] is called.
-  final Reducible<S> reducible;
+  final ReducedStore<S> store;
 
-  /// The reducer that is passed as a parameter to the [reduce](Reducible.reduce) method
-  /// of the [reducible] when the [call] method is called.
+  /// The reducer that is passed as a parameter to the [reduce](ReducedStore.reduce) method
+  /// of the [store] when the [call] method is called.
   final Reducer1<S, V> reducer;
 
-  /// Executes the [reduce](Reducible.reduce) method of the [reducible]
+  /// Executes the [reduce](ReducedStore.reduce) method of the [store]
   ///  with the [reducer] as parameter.
   @override
-  call(value) => reducible.reduce(Reducer1Adapter(reducer, value));
+  call(value) => store.reduce(Reducer1Adapter(reducer, value));
 
   /// For this class to have value semantics, both constructor parameters
-  /// [reducible] and [reducer] should have value semantics.
+  /// [store] and [reducer] should have value semantics.
   @override
-  get hashCode => Object.hash(reducible, reducer);
+  get hashCode => Object.hash(store, reducer);
 
   /// For this class to have value semantics, both constructor parameters
-  /// [reducible] and [reducer] should have value semantics.
+  /// [store] and [reducer] should have value semantics.
   @override
   operator ==(other) =>
       other is Callable1Adapter &&
       reducer == other.reducer &&
-      reducible == other.reducible;
+      store == other.store;
 }
 
-/// An implementation of a callback as a [Reducible.reduce](Reducible.reduce) call with a [Reducer2].
+/// An implementation of a callback as a [ReducedStore.reduce](ReducedStore.reduce) call with a [Reducer2].
 ///
-/// Or in other words, a [Reducer2] bonded to a [Reducible] useable as callback.
+/// Or in other words, a [Reducer2] bonded to a [ReducedStore] useable as callback.
 ///
-/// The type parameter `S` is the type of the state of the [Reducible].
+/// The type parameter `S` is the type of the state of the [ReducedStore].
 /// The type parameter `V1` is the type of the 1st value of the [Reducer2].
 /// The type parameter `V2` is the type of the 2nd value of the [Reducer2].
 class Callable2Adapter<S, V1, V2> extends Callable2<void, V1, V2> {
-  const Callable2Adapter(this.reducible, this.reducer);
+  const Callable2Adapter(this.store, this.reducer);
 
-  /// The reducible to whose method [reduce](Reducible.reduce)
+  /// The store to whose method [reduce](ReducedStore.reduce)
   /// the [reducer] is passed when the method [call] is called.
-  final Reducible<S> reducible;
+  final ReducedStore<S> store;
 
-  /// The reducer that is passed as a parameter to the [reduce](Reducible.reduce) method
-  /// of the [reducible] when the [call] method is called.
+  /// The reducer that is passed as a parameter to the [reduce](ReducedStore.reduce) method
+  /// of the [store] when the [call] method is called.
   final Reducer2<S, V1, V2> reducer;
 
-  /// Executes the [reduce](Reducible.reduce) method of the [reducible]
+  /// Executes the [reduce](ReducedStore.reduce) method of the [store]
   ///  with the [reducer] as parameter.
   @override
   call(value1, value2) =>
-      reducible.reduce(Reducer2Adapter(reducer, value1, value2));
+      store.reduce(Reducer2Adapter(reducer, value1, value2));
 
   /// For this class to have value semantics, both constructor parameters
-  /// [reducible] and [reducer] should have value semantics.
+  /// [store] and [reducer] should have value semantics.
   @override
-  get hashCode => Object.hash(reducible, reducer);
+  get hashCode => Object.hash(store, reducer);
 
   /// For this class to have value semantics, both constructor parameters
-  /// [reducible] and [reducer] should have value semantics.
+  /// [store] and [reducer] should have value semantics.
   @override
   operator ==(other) =>
       other is Callable2Adapter &&
       reducer == other.reducer &&
-      reducible == other.reducible;
+      store == other.store;
 }
 
-/// An implementation of a callback as a [Reducible.reduce](Reducible.reduce) call with a [Reducer3].
+/// An implementation of a callback as a [ReducedStore.reduce](ReducedStore.reduce) call with a [Reducer3].
 ///
-/// Or in other words, a [Reducer3] bonded to a [Reducible] useable as callback.
+/// Or in other words, a [Reducer3] bonded to a [ReducedStore] useable as callback.
 ///
-/// The type parameter `S` is the type of the state of the [Reducible].
+/// The type parameter `S` is the type of the state of the [ReducedStore].
 /// The type parameter `V1` is the type of the 1st value of the [Reducer3].
 /// The type parameter `V2` is the type of the 2nd value of the [Reducer3].
 /// The type parameter `V3` is the type of the 3rd value of the [Reducer3].
-class Callable3Adapter<S, V1, V2, V3> extends Callable3<void, V1, V2, V3> {
-  const Callable3Adapter(this.reducible, this.reducer);
+class Callable3Adapter<S, V1, V2, V3>
+    extends Callable3<void, V1, V2, V3> {
+  const Callable3Adapter(this.store, this.reducer);
 
-  /// The reducible to whose method [reduce](Reducible.reduce)
+  /// The store to whose method [reduce](ReducedStore.reduce)
   /// the [reducer] is passed when the method [call] is called.
-  final Reducible<S> reducible;
+  final ReducedStore<S> store;
 
-  /// The reducer that is passed as a parameter to the [reduce](Reducible.reduce) method
-  /// of the [reducible] when the [call] method is called.
+  /// The reducer that is passed as a parameter to the [reduce](ReducedStore.reduce) method
+  /// of the [store] when the [call] method is called.
   final Reducer3<S, V1, V2, V3> reducer;
 
-  /// Executes the [reduce](Reducible.reduce) method of the [reducible]
+  /// Executes the [reduce](ReducedStore.reduce) method of the [store]
   ///  with the [reducer] as parameter.
   @override
   call(value1, value2, value3) =>
-      reducible.reduce(Reducer3Adapter(reducer, value1, value2, value3));
+      store.reduce(Reducer3Adapter(reducer, value1, value2, value3));
 
   /// For this class to have value semantics, both constructor parameters
-  /// [reducible] and [reducer] should have value semantics.
+  /// [store] and [reducer] should have value semantics.
   @override
-  get hashCode => Object.hash(reducible, reducer);
+  get hashCode => Object.hash(store, reducer);
 
   /// For this class to have value semantics, both constructor parameters
-  /// [reducible] and [reducer] should have value semantics.
+  /// [store] and [reducer] should have value semantics.
   @override
   operator ==(other) =>
       other is Callable3Adapter &&
       reducer == other.reducer &&
-      reducible == other.reducible;
+      store == other.store;
 }