+++ test/callable/callable1_adapter_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:reduced/src/callable.dart';
-import 'package:reduced/src/reducible.dart';
+import 'package:reduced/src/reducer.dart';
+import 'package:reduced/src/store.dart';
 
 class MockReducer1 extends Reducer1<String, String> {
   MockReducer1();
@@ -9,8 +10,8 @@ class MockReducer1 extends Reducer1<String, String> {
   call(state, value) => '$state $value';
 }
 
-class MockReducible extends Reducible<String> {
-  MockReducible(this.state);
+class MockReducedStore extends ReducedStore<String> {
+  MockReducedStore(this.state);
 
   @override
   String state;
@@ -22,66 +23,69 @@ class MockReducible extends Reducible<String> {
 void main() {
   test('Callable1Adapter init test', () {
     final objectUnterTest = Callable1Adapter(
-      MockReducible('0'),
+      MockReducedStore('0'),
       MockReducer1(),
     );
-    expect(objectUnterTest.reducible.state, '0');
+    expect(objectUnterTest.store.state, '0');
   });
 
   test('Callable1Adapter call test', () {
     final objectUnterTest = Callable1Adapter(
-      MockReducible('0'),
+      MockReducedStore('0'),
       MockReducer1(),
     );
     objectUnterTest.call('1');
-    expect(objectUnterTest.reducible.state, '0 1');
+    expect(objectUnterTest.store.state, '0 1');
   });
 
   test('Callable1Adapter hashCode test', () {
-    final reducible1 = MockReducible('1');
+    final store1 = MockReducedStore('1');
     final reducer1 = MockReducer1();
-    final reducible2 = MockReducible('2');
+    final store2 = MockReducedStore('2');
     final reducer2 = MockReducer1();
     final objectUnterTest11 = Callable1Adapter(
-      reducible1,
+      store1,
       reducer1,
     );
     final objectUnterTest12 = Callable1Adapter(
-      reducible1,
+      store1,
       reducer2,
     );
     final objectUnterTest21 = Callable1Adapter(
-      reducible2,
+      store2,
       reducer1,
     );
     final objectUnterTest22 = Callable1Adapter(
-      reducible1,
+      store1,
       reducer1,
     );
     expect(objectUnterTest11.hashCode, objectUnterTest22.hashCode);
-    expect(objectUnterTest11.hashCode, isNot(objectUnterTest12.hashCode));
-    expect(objectUnterTest11.hashCode, isNot(objectUnterTest21.hashCode));
-    expect(objectUnterTest12.hashCode, isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest11.hashCode,
+        isNot(objectUnterTest12.hashCode));
+    expect(objectUnterTest11.hashCode,
+        isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest12.hashCode,
+        isNot(objectUnterTest21.hashCode));
   });
   test('Callable1Adapter operator== test', () {
-    final reducible1 = MockReducible('1');
+    final store1 = MockReducedStore('1');
     final reducer1 = MockReducer1();
-    final reducible2 = MockReducible('2');
+    final store2 = MockReducedStore('2');
     final reducer2 = MockReducer1();
     final objectUnterTest11 = Callable1Adapter(
-      reducible1,
+      store1,
       reducer1,
     );
     final objectUnterTest12 = Callable1Adapter(
-      reducible1,
+      store1,
       reducer2,
     );
     final objectUnterTest21 = Callable1Adapter(
-      reducible2,
+      store2,
       reducer1,
     );
     final objectUnterTest22 = Callable1Adapter(
-      reducible1,
+      store1,
       reducer1,
     );
     expect(objectUnterTest11 == objectUnterTest22, true);