+++ test/callable/callable2_adapter_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:reduced/src/callable.dart';
-import 'package:reduced/src/reducible.dart';
+import 'package:reduced/src/reducer.dart';
+import 'package:reduced/src/store.dart';
 
 class MockReducer2 extends Reducer2<String, String, String> {
   MockReducer2();
@@ -9,8 +10,8 @@ class MockReducer2 extends Reducer2<String, String, String> {
   call(state, value1, value2) => '$state $value1 $value2';
 }
 
-class MockReducible extends Reducible<String> {
-  MockReducible(this.state);
+class MockReducedStore extends ReducedStore<String> {
+  MockReducedStore(this.state);
 
   @override
   String state;
@@ -22,66 +23,69 @@ class MockReducible extends Reducible<String> {
 void main() {
   test('Callable2Adapter init test', () {
     final objectUnterTest = Callable2Adapter(
-      MockReducible('0'),
+      MockReducedStore('0'),
       MockReducer2(),
     );
-    expect(objectUnterTest.reducible.state, '0');
+    expect(objectUnterTest.store.state, '0');
   });
 
   test('Callable2Adapter call test', () {
     final objectUnterTest = Callable2Adapter(
-      MockReducible('0'),
+      MockReducedStore('0'),
       MockReducer2(),
     );
     objectUnterTest.call('1', '2');
-    expect(objectUnterTest.reducible.state, '0 1 2');
+    expect(objectUnterTest.store.state, '0 1 2');
   });
 
   test('Callable2Adapter hashCode test', () {
-    final reducible1 = MockReducible('1');
+    final store1 = MockReducedStore('1');
     final reducer1 = MockReducer2();
-    final reducible2 = MockReducible('2');
+    final store2 = MockReducedStore('2');
     final reducer2 = MockReducer2();
     final objectUnterTest11 = Callable2Adapter(
-      reducible1,
+      store1,
       reducer1,
     );
     final objectUnterTest12 = Callable2Adapter(
-      reducible1,
+      store1,
       reducer2,
     );
     final objectUnterTest21 = Callable2Adapter(
-      reducible2,
+      store2,
       reducer1,
     );
     final objectUnterTest22 = Callable2Adapter(
-      reducible1,
+      store1,
       reducer1,
     );
     expect(objectUnterTest11.hashCode, objectUnterTest22.hashCode);
-    expect(objectUnterTest11.hashCode, isNot(objectUnterTest12.hashCode));
-    expect(objectUnterTest11.hashCode, isNot(objectUnterTest21.hashCode));
-    expect(objectUnterTest12.hashCode, isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest11.hashCode,
+        isNot(objectUnterTest12.hashCode));
+    expect(objectUnterTest11.hashCode,
+        isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest12.hashCode,
+        isNot(objectUnterTest21.hashCode));
   });
   test('Callable2Adapter operator== test', () {
-    final reducible1 = MockReducible('1');
+    final store1 = MockReducedStore('1');
     final reducer1 = MockReducer2();
-    final reducible2 = MockReducible('2');
+    final store2 = MockReducedStore('2');
     final reducer2 = MockReducer2();
     final objectUnterTest11 = Callable2Adapter(
-      reducible1,
+      store1,
       reducer1,
     );
     final objectUnterTest12 = Callable2Adapter(
-      reducible1,
+      store1,
       reducer2,
     );
     final objectUnterTest21 = Callable2Adapter(
-      reducible2,
+      store2,
       reducer1,
     );
     final objectUnterTest22 = Callable2Adapter(
-      reducible1,
+      store1,
       reducer1,
     );
     expect(objectUnterTest11 == objectUnterTest22, true);