+++ test/callable/callable_adapter_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:reduced/src/callable.dart';
-import 'package:reduced/src/reducible.dart';
+import 'package:reduced/src/reducer.dart';
+import 'package:reduced/src/store.dart';
 
 class MockReducer extends Reducer<Object> {
   MockReducer(this.newState);
@@ -11,8 +12,8 @@ class MockReducer extends Reducer<Object> {
   call(state) => newState;
 }
 
-class MockReducible extends Reducible<Object> {
-  MockReducible(this.state);
+class MockReducedStore extends ReducedStore<Object> {
+  MockReducedStore(this.state);
 
   @override
   Object state;
@@ -24,66 +25,69 @@ class MockReducible extends Reducible<Object> {
 void main() {
   test('CallableAdapter init test', () {
     final objectUnterTest = CallableAdapter(
-      MockReducible(0),
+      MockReducedStore(0),
       MockReducer(1),
     );
-    expect(objectUnterTest.reducible.state, 0);
+    expect(objectUnterTest.store.state, 0);
   });
 
   test('CallableAdapter call test', () {
     final objectUnterTest = CallableAdapter(
-      MockReducible(0),
+      MockReducedStore(0),
       MockReducer(1),
     );
     objectUnterTest.call();
-    expect(objectUnterTest.reducible.state, 1);
+    expect(objectUnterTest.store.state, 1);
   });
 
   test('CallableAdapter hashCode test', () {
-    final reducible1 = MockReducible(1);
+    final store1 = MockReducedStore(1);
     final reducer1 = MockReducer(1);
-    final reducible2 = MockReducible(2);
+    final store2 = MockReducedStore(2);
     final reducer2 = MockReducer(2);
     final objectUnterTest11 = CallableAdapter(
-      reducible1,
+      store1,
       reducer1,
     );
     final objectUnterTest12 = CallableAdapter(
-      reducible1,
+      store1,
       reducer2,
     );
     final objectUnterTest21 = CallableAdapter(
-      reducible2,
+      store2,
       reducer1,
     );
     final objectUnterTest22 = CallableAdapter(
-      reducible1,
+      store1,
       reducer1,
     );
     expect(objectUnterTest11.hashCode, objectUnterTest22.hashCode);
-    expect(objectUnterTest11.hashCode, isNot(objectUnterTest12.hashCode));
-    expect(objectUnterTest11.hashCode, isNot(objectUnterTest21.hashCode));
-    expect(objectUnterTest12.hashCode, isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest11.hashCode,
+        isNot(objectUnterTest12.hashCode));
+    expect(objectUnterTest11.hashCode,
+        isNot(objectUnterTest21.hashCode));
+    expect(objectUnterTest12.hashCode,
+        isNot(objectUnterTest21.hashCode));
   });
   test('CallableAdapter operator== test', () {
-    final reducible1 = MockReducible(1);
+    final store1 = MockReducedStore(1);
     final reducer1 = MockReducer(1);
-    final reducible2 = MockReducible(2);
+    final store2 = MockReducedStore(2);
     final reducer2 = MockReducer(2);
     final objectUnterTest11 = CallableAdapter(
-      reducible1,
+      store1,
       reducer1,
     );
     final objectUnterTest12 = CallableAdapter(
-      reducible1,
+      store1,
       reducer2,
     );
     final objectUnterTest21 = CallableAdapter(
-      reducible2,
+      store2,
       reducer1,
     );
     final objectUnterTest22 = CallableAdapter(
-      reducible1,
+      store1,
       reducer1,
     );
     expect(objectUnterTest11 == objectUnterTest22, true);